+++ src/vita/modulesHandler.cpp
@@ -1,5 +1,5 @@
 #include "modulesHandler.h"
-bool detectInbound(int x1,int y1,int width1,int heigth1,int x2,int y2,int width2,int heigth2){
+bool detectInbound(float x1,float y1,float width1,float heigth1,float x2,float y2,float width2,float heigth2){
     return !(
         x1+width1<x2 ||
         x1>x2+width2 ||
